[Hotfix][CDC] Fix package name spelling mistake (#7415)
