+++ seatunnel-connectors-v2/connector-cdc/connector-cdc-oracle/src/main/java/org/apache/seatunnel/connectors/seatunnel/cdc/oracle/source/enumerator/OracleChunkSplitter.java
@@ -15,50 +15,58 @@
  * limitations under the License.
  */
 
-package org.apache.seatunnel.connectors.seatunnel.cdc.sqlserver.source.source.eumerator;
+package org.apache.seatunnel.connectors.seatunnel.cdc.oracle.source.enumerator;
 
 import org.apache.seatunnel.api.table.type.SeaTunnelDataType;
 import org.apache.seatunnel.api.table.type.SeaTunnelRowType;
 import org.apache.seatunnel.connectors.cdc.base.config.JdbcSourceConfig;
 import org.apache.seatunnel.connectors.cdc.base.dialect.JdbcDataSourceDialect;
 import org.apache.seatunnel.connectors.cdc.base.source.enumerator.splitter.AbstractJdbcSourceChunkSplitter;
-import org.apache.seatunnel.connectors.seatunnel.cdc.sqlserver.source.utils.SqlServerTypeUtils;
-import org.apache.seatunnel.connectors.seatunnel.cdc.sqlserver.source.utils.SqlServerUtils;
+import org.apache.seatunnel.connectors.cdc.base.utils.ObjectUtils;
+import org.apache.seatunnel.connectors.seatunnel.cdc.oracle.config.OracleSourceConfig;
+import org.apache.seatunnel.connectors.seatunnel.cdc.oracle.utils.OracleTypeUtils;
+import org.apache.seatunnel.connectors.seatunnel.cdc.oracle.utils.OracleUtils;
 
 import io.debezium.jdbc.JdbcConnection;
 import io.debezium.relational.Column;
 import io.debezium.relational.Table;
 import io.debezium.relational.TableId;
 import lombok.extern.slf4j.Slf4j;
+import oracle.sql.ROWID;
 
 import java.sql.SQLException;
 
-/** The {@code ChunkSplitter} used to split table into a set of chunks for JDBC data source. */
+/**
+ * The {@code ChunkSplitter} used to split Oracle table into a set of chunks for JDBC data source.
+ */
 @Slf4j
-public class SqlServerChunkSplitter extends AbstractJdbcSourceChunkSplitter {
+public class OracleChunkSplitter extends AbstractJdbcSourceChunkSplitter {
 
-    public SqlServerChunkSplitter(JdbcSourceConfig sourceConfig, JdbcDataSourceDialect dialect) {
+    private final OracleSourceConfig oracleSourceConfig;
+
+    public OracleChunkSplitter(JdbcSourceConfig sourceConfig, JdbcDataSourceDialect dialect) {
         super(sourceConfig, dialect);
+        this.oracleSourceConfig = (OracleSourceConfig) sourceConfig;
     }
 
     @Override
     public Object[] queryMinMax(JdbcConnection jdbc, TableId tableId, String columnName)
             throws SQLException {
-        return SqlServerUtils.queryMinMax(jdbc, tableId, columnName);
+        return OracleUtils.queryMinMax(jdbc, tableId, columnName);
     }
 
     @Override
     public Object queryMin(
             JdbcConnection jdbc, TableId tableId, String columnName, Object excludedLowerBound)
             throws SQLException {
-        return SqlServerUtils.queryMin(jdbc, tableId, columnName, excludedLowerBound);
+        return OracleUtils.queryMin(jdbc, tableId, columnName, excludedLowerBound);
     }
 
     @Override
     public Object[] sampleDataFromColumn(
             JdbcConnection jdbc, TableId tableId, String columnName, int inverseSamplingRate)
             throws Exception {
-        return SqlServerUtils.skipReadAndSortSampleData(
+        return OracleUtils.skipReadAndSortSampleData(
                 jdbc, tableId, columnName, inverseSamplingRate);
     }
 
@@ -70,24 +78,31 @@ public class SqlServerChunkSplitter extends AbstractJdbcSourceChunkSplitter {
             int chunkSize,
             Object includedLowerBound)
             throws SQLException {
-        return SqlServerUtils.queryNextChunkMax(
+        return OracleUtils.queryNextChunkMax(
                 jdbc, tableId, columnName, chunkSize, includedLowerBound);
     }
 
     @Override
     public Long queryApproximateRowCnt(JdbcConnection jdbc, TableId tableId) throws SQLException {
-        return SqlServerUtils.queryApproximateRowCnt(jdbc, tableId);
+        return OracleUtils.queryApproximateRowCnt(oracleSourceConfig, jdbc, tableId);
     }
 
     @Override
     public String buildSplitScanQuery(
             Table table, SeaTunnelRowType splitKeyType, boolean isFirstSplit, boolean isLastSplit) {
-        return SqlServerUtils.buildSplitScanQuery(
-                table.id(), splitKeyType, isFirstSplit, isLastSplit);
+        return OracleUtils.buildSplitScanQuery(table.id(), splitKeyType, isFirstSplit, isLastSplit);
     }
 
     @Override
     public SeaTunnelDataType<?> fromDbzColumn(Column splitColumn) {
-        return SqlServerTypeUtils.convertFromColumn(splitColumn);
+        return OracleTypeUtils.convertFromColumn(splitColumn);
+    }
+
+    protected int ObjectCompare(Object obj1, Object obj2) {
+        if (obj1 instanceof ROWID && obj2 instanceof ROWID) {
+            return ROWID.compareBytes(((ROWID) obj1).getBytes(), ((ROWID) obj2).getBytes());
+        } else {
+            return ObjectUtils.compare(obj1, obj2);
+        }
     }
 }
