+++ seatunnel-connectors-v2/connector-cdc/connector-cdc-sqlserver/src/main/java/org/apache/seatunnel/connectors/seatunnel/cdc/sqlserver/source/source/enumerator/SqlServerChunkSplitter.java
@@ -15,58 +15,50 @@
  * limitations under the License.
  */
 
-package org.apache.seatunnel.connectors.seatunnel.cdc.oracle.source.eumerator;
+package org.apache.seatunnel.connectors.seatunnel.cdc.sqlserver.source.source.enumerator;
 
 import org.apache.seatunnel.api.table.type.SeaTunnelDataType;
 import org.apache.seatunnel.api.table.type.SeaTunnelRowType;
 import org.apache.seatunnel.connectors.cdc.base.config.JdbcSourceConfig;
 import org.apache.seatunnel.connectors.cdc.base.dialect.JdbcDataSourceDialect;
 import org.apache.seatunnel.connectors.cdc.base.source.enumerator.splitter.AbstractJdbcSourceChunkSplitter;
-import org.apache.seatunnel.connectors.cdc.base.utils.ObjectUtils;
-import org.apache.seatunnel.connectors.seatunnel.cdc.oracle.config.OracleSourceConfig;
-import org.apache.seatunnel.connectors.seatunnel.cdc.oracle.utils.OracleTypeUtils;
-import org.apache.seatunnel.connectors.seatunnel.cdc.oracle.utils.OracleUtils;
+import org.apache.seatunnel.connectors.seatunnel.cdc.sqlserver.source.utils.SqlServerTypeUtils;
+import org.apache.seatunnel.connectors.seatunnel.cdc.sqlserver.source.utils.SqlServerUtils;
 
 import io.debezium.jdbc.JdbcConnection;
 import io.debezium.relational.Column;
 import io.debezium.relational.Table;
 import io.debezium.relational.TableId;
 import lombok.extern.slf4j.Slf4j;
-import oracle.sql.ROWID;
 
 import java.sql.SQLException;
 
-/**
- * The {@code ChunkSplitter} used to split Oracle table into a set of chunks for JDBC data source.
- */
+/** The {@code ChunkSplitter} used to split table into a set of chunks for JDBC data source. */
 @Slf4j
-public class OracleChunkSplitter extends AbstractJdbcSourceChunkSplitter {
+public class SqlServerChunkSplitter extends AbstractJdbcSourceChunkSplitter {
 
-    private final OracleSourceConfig oracleSourceConfig;
-
-    public OracleChunkSplitter(JdbcSourceConfig sourceConfig, JdbcDataSourceDialect dialect) {
+    public SqlServerChunkSplitter(JdbcSourceConfig sourceConfig, JdbcDataSourceDialect dialect) {
         super(sourceConfig, dialect);
-        this.oracleSourceConfig = (OracleSourceConfig) sourceConfig;
     }
 
     @Override
     public Object[] queryMinMax(JdbcConnection jdbc, TableId tableId, String columnName)
             throws SQLException {
-        return OracleUtils.queryMinMax(jdbc, tableId, columnName);
+        return SqlServerUtils.queryMinMax(jdbc, tableId, columnName);
     }
 
     @Override
     public Object queryMin(
             JdbcConnection jdbc, TableId tableId, String columnName, Object excludedLowerBound)
             throws SQLException {
-        return OracleUtils.queryMin(jdbc, tableId, columnName, excludedLowerBound);
+        return SqlServerUtils.queryMin(jdbc, tableId, columnName, excludedLowerBound);
     }
 
     @Override
     public Object[] sampleDataFromColumn(
             JdbcConnection jdbc, TableId tableId, String columnName, int inverseSamplingRate)
             throws Exception {
-        return OracleUtils.skipReadAndSortSampleData(
+        return SqlServerUtils.skipReadAndSortSampleData(
                 jdbc, tableId, columnName, inverseSamplingRate);
     }
 
@@ -78,31 +70,24 @@ public class OracleChunkSplitter extends AbstractJdbcSourceChunkSplitter {
             int chunkSize,
             Object includedLowerBound)
             throws SQLException {
-        return OracleUtils.queryNextChunkMax(
+        return SqlServerUtils.queryNextChunkMax(
                 jdbc, tableId, columnName, chunkSize, includedLowerBound);
     }
 
     @Override
     public Long queryApproximateRowCnt(JdbcConnection jdbc, TableId tableId) throws SQLException {
-        return OracleUtils.queryApproximateRowCnt(oracleSourceConfig, jdbc, tableId);
+        return SqlServerUtils.queryApproximateRowCnt(jdbc, tableId);
     }
 
     @Override
     public String buildSplitScanQuery(
             Table table, SeaTunnelRowType splitKeyType, boolean isFirstSplit, boolean isLastSplit) {
-        return OracleUtils.buildSplitScanQuery(table.id(), splitKeyType, isFirstSplit, isLastSplit);
+        return SqlServerUtils.buildSplitScanQuery(
+                table.id(), splitKeyType, isFirstSplit, isLastSplit);
     }
 
     @Override
     public SeaTunnelDataType<?> fromDbzColumn(Column splitColumn) {
-        return OracleTypeUtils.convertFromColumn(splitColumn);
-    }
-
-    protected int ObjectCompare(Object obj1, Object obj2) {
-        if (obj1 instanceof ROWID && obj2 instanceof ROWID) {
-            return ROWID.compareBytes(((ROWID) obj1).getBytes(), ((ROWID) obj2).getBytes());
-        } else {
-            return ObjectUtils.compare(obj1, obj2);
-        }
+        return SqlServerTypeUtils.convertFromColumn(splitColumn);
     }
 }
